Adding a change log for runs and recipes

+++ packages/catts_engine/src/chain_config/state.rs
@@ -1,7 +1,7 @@
 use super::{ChainConfig, ChainConfigError};
 use crate::CHAIN_CONFIGS;
 
-pub fn get(chain_id: u64) -> Result<ChainConfig, ChainConfigError> {
+pub fn get(chain_id: u32) -> Result<ChainConfig, ChainConfigError> {
     CHAIN_CONFIGS
         .with_borrow(|configs| configs.get(&chain_id))
         .ok_or(ChainConfigError::NotFound)
